+++ app/src/main/java/com/vci/vectorcamapp/MainActivity.kt
@@ -29,7 +29,7 @@ import com.vci.vectorcamapp.core.presentation.util.error.LocalErrorMessageEmitte
 import com.vci.vectorcamapp.navigation.NavGraph
 import com.vci.vectorcamapp.ui.theme.VectorcamappTheme
 import dagger.hilt.android.AndroidEntryPoint
-import com.vci.vectorcamapp.core.domain.util.ErrorMessageEmitter
+import com.vci.vectorcamapp.core.presentation.util.error.ErrorMessageEmitter
 import javax.inject.Inject
 
 @AndroidEntryPoint
