+++ app/src/main/java/com/vci/vectorcamapp/core/di/ErrorMessageModule.kt
@@ -1,7 +1,7 @@
 package com.vci.vectorcamapp.core.di
 
-import com.vci.vectorcamapp.core.data.DefaultErrorMessageEmitter
-import com.vci.vectorcamapp.core.domain.util.ErrorMessageEmitter
+import com.vci.vectorcamapp.core.presentation.util.error.DefaultErrorMessageEmitter
+import com.vci.vectorcamapp.core.presentation.util.error.ErrorMessageEmitter
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
