+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/util/error/DefaultErrorMessageEmitter.kt
@@ -1,9 +1,8 @@
-package com.vci.vectorcamapp.core.data
+package com.vci.vectorcamapp.core.presentation.util.error
 
 import androidx.compose.material3.SnackbarDuration
 import com.vci.vectorcamapp.core.domain.util.Error
 import com.vci.vectorcamapp.core.presentation.model.ErrorData
-import com.vci.vectorcamapp.core.domain.util.ErrorMessageEmitter
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
