+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/util/error/ErrorMessageEmitter.kt
@@ -1,6 +1,7 @@
-package com.vci.vectorcamapp.core.domain.util
+package com.vci.vectorcamapp.core.presentation.util.error
 
 import androidx.compose.material3.SnackbarDuration
+import com.vci.vectorcamapp.core.domain.util.Error
 import com.vci.vectorcamapp.core.presentation.model.ErrorData
 import kotlinx.coroutines.flow.SharedFlow
