+++ app/src/main/java/com/vci/vectorcamapp/landing/presentation/LandingViewModel.kt
@@ -7,7 +7,7 @@ import com.vci.vectorcamapp.core.domain.model.enums.SessionType
 import com.vci.vectorcamapp.core.domain.repository.ProgramRepository
 import com.vci.vectorcamapp.core.domain.repository.SessionRepository
 import com.vci.vectorcamapp.core.presentation.CoreViewModel
-import com.vci.vectorcamapp.core.domain.util.ErrorMessageEmitter
+import com.vci.vectorcamapp.core.presentation.util.error.ErrorMessageEmitter
 import com.vci.vectorcamapp.landing.domain.util.LandingError
 import com.vci.vectorcamapp.landing.logging.LandingSentryLogger
 import dagger.hilt.android.lifecycle.HiltViewModel
